grabber & actuator changes
Actuator Subsystem class is deleted, along with ActuatorUp & ActuatorDown Command classes
Rewrote Climber subsystem class so that the climber only use 1 motor.
In the Grabber Subsystem class, the open() and close() functions are changed. Added an if else statement to close(). Both functions are written with the assumption that the limit switch, when not pressed, returns True, and when pressed, returns False.