changed how Arm, Rotator, & Climber subsystems are controlled
Rotator used to be controlled by right stick. This is changed so that the Rotator is controlled by 2 Xbox buttons (A & Y).
Arm subsystem is changed to be controlled by triggers, with right trigger to raise arm, left trigger to lower arm
Climber is changed to use the right stick to raise actuator & climb. A command that uses Climber is made to allow to robot to maintain a specific height after it climbed.