xbox buttons & autonomous changes
Robot controlling is now done using the Xbox only. The joystick has no role in controlling the robot. The buttons to control the different subsystems are now on the Xbox rather than the joystick.
DriveSubsystem is changed to use the Xbox left stick.
In ClimberMove, the argument for climber.move() is given a negative sign. This should allow the actuator to extend when the stick is pushed forward.
In the autonomous commands, all subsystems other than the drive are commented out. So the robot only drives during autonomous. The arm, the rotator, the grabber, and the climber won't move during autonomous. This is to simplify the autonomous code so that the robot only crosses the line during autonomous.